+++ src/common/base/base.controller.ts
@@ -1,4 +1,4 @@
-import { ApiNotFoundResponse, ApiOperation } from '@nestjs/swagger';
+import { ApiExcludeEndpoint, ApiNotFoundResponse, ApiOperation } from '@nestjs/swagger';
 import { CacheInterceptor, Get, NotFoundException, Param, Query, UseInterceptors, Version } from '@nestjs/common';
 import { IQuery } from '../interfaces/query.interface';
 import { Paginated } from '../decorators/paginated.decorator';
@@ -11,15 +11,17 @@ export function BaseController<TEntity, TEntityDto>(
   Entity: Constructor<TEntity>,
   EntityDto: Constructor<TEntityDto>,
   description?: string,
+  version?: string,
 ) {
   abstract class BaseController {
     protected constructor(readonly service: any) {}
 
-    @Version('1')
+    @Version(version ?? '1')
     @Get()
     @ApiOperation({ summary: description })
     @UseInterceptors(CacheInterceptor)
     @Paginated(EntityDto, Entity, { findForAllProperties: true })
+    @ApiExcludeEndpoint()
     async findManyByQuery(@Query() query: IQuery): Promise<TEntityDto> {
       return this.service.findMany(query);
     }
@@ -33,25 +35,28 @@ export function BaseControllerWithFindById<TEntity, TEntityDto>(
   EntityDto: Constructor<TEntityDto>,
   summary?: string,
   description?: string,
+  version?: string,
 ) {
   abstract class BaseControllerWithFindById {
     protected constructor(readonly service: any) {}
 
-    @Version('1')
+    @Version(version ?? '1')
     @Get()
     @UseInterceptors(CacheInterceptor)
     @ApiOperation({ summary, description })
     @Paginated(EntityDto, Entity, { findForAllProperties: true })
+    @ApiExcludeEndpoint()
     async findManyByQuery(@Query() query: IQuery): Promise<TEntityDto> {
       return this.service.findMany(query);
     }
 
-    @Version('1')
+    @Version(version ?? '1')
     @Get(':id')
     @UseInterceptors(CacheInterceptor)
     @ApiOperation({ summary: 'Поиск по id', description: 'Возвращает всю доступную информацию о сущности.' })
     @ApiBaseResponse({ type: Entity })
     @ApiNotFoundResponse({ type: ForbiddenErrorResponseDto, description: 'NotFound' })
+    @ApiExcludeEndpoint()
     async findOne(@Param('id') id: string): Promise<TEntity> {
       const found = await this.service.findOne(+id);
       if (!found) throw new NotFoundException('По этому id ничего не найдено!');