+++ src/common/decorators/transform/to-array.decorator.ts
@@ -1,0 +1,9 @@
+import { Transform } from 'class-transformer';
+
+export const ToArray = () => {
+  return Transform(({ value }: { value: unknown }) => {
+    if (value === undefined) return value;
+    if (Array.isArray(value)) return value;
+    return [value];
+  });
+};