+++ src/person/dto/v1.4/person-docs.response.ts
@@ -1,0 +1,8 @@
+import { PersonDocsResponseDto } from '../person-docs.response';
+
+export class PersonDocsResponseDtoV1_4 extends PersonDocsResponseDto {
+  constructor(partial: Partial<PersonDocsResponseDtoV1_4>) {
+    super(partial);
+    Object.assign(this, partial);
+  }
+}