+++ src/search-sync/search-sync.service.ts
@@ -3,8 +3,8 @@ import { Cron, CronExpression } from '@nestjs/schedule';
 import { MeiliService } from 'src/meili/meili.service';
 import { MovieService } from 'src/movie/movie.service';
 import { PersonService } from 'src/person/person.service';
-import { MOVIE_INDEX } from '../movie/constants/movie-index';
-import { PERSON_INDEX } from '../person/constants/person-index';
+import { MOVIE_INDEX, MOVIE_V1_4_INDEX } from '../movie/constants/movie-index';
+import { PERSON_INDEX, PERSON_V1_4_INDEX } from '../person/constants/person-index';
 import { Person } from '../person/schemas/person.schema';
 import { Movie } from '../movie/schemas/movie.schema';
 import { MeiliMovieEntity } from '../movie/entities/meili-movie.entity';
@@ -12,8 +12,10 @@ import { MeiliPersonEntity } from '../person/entities/meili-person.entity';
 import { Model } from 'mongoose';
 import { InjectModel } from '@nestjs/mongoose';
 import { SearchSync, SearchSyncDocument } from './schemas/search-sync.schema';
+import { MeiliMovieEntityV1_4 } from '../movie/entities/v1.4/meili-movie.entity';
+import { MeiliPersonEntityV1_4 } from '../person/entities/v1.4/meili-person.entity';
 
-type EntityTypes = typeof MOVIE_INDEX | typeof PERSON_INDEX;
+type EntityTypes = typeof MOVIE_INDEX | typeof PERSON_INDEX | typeof MOVIE_V1_4_INDEX | typeof PERSON_V1_4_INDEX;
 
 @Injectable()
 export class SearchSyncService implements OnModuleInit {
@@ -34,12 +36,9 @@ export class SearchSyncService implements OnModuleInit {
     }
   }
 
-  private async syncEntity<Entity>(
-    entityType: EntityTypes,
-    service: MovieService | PersonService,
-    pageSize = 1000,
-  ): Promise<void> {
+  private async syncEntity<Entity>(entityType: EntityTypes, service: MovieService | PersonService, pageSize = 1000): Promise<void> {
     return new Promise(async (resolve) => {
+      this.logger.log(`Starting sync for ${entityType}`);
       const process = await this.searchSyncModel.findOne({ entityType });
       if (process?.processing) return;
       await this.searchSyncModel.updateOne({ entityType }, { processing: true }, { upsert: true });
@@ -55,13 +54,20 @@ export class SearchSyncService implements OnModuleInit {
         };
 
         const result = await service.findMany(query);
+        this.logger.log(`Processing. Page ${pageIndex} - ${result.docs.length} items`);
         let entities = [];
         switch (entityType) {
           case MOVIE_INDEX:
             entities = (result.docs as Movie[]).map((movie) => new MeiliMovieEntity({}).fromMongoDocument(movie));
             break;
+          case MOVIE_V1_4_INDEX:
+            entities = (result.docs as Movie[]).map((movie) => new MeiliMovieEntityV1_4({}).fromMongoDocument(movie));
+            break;
           case PERSON_INDEX:
             entities = (result.docs as Person[]).map((person) => new MeiliPersonEntity({}).fromMongoDocument(person));
+            break;
+          case PERSON_V1_4_INDEX:
+            entities = (result.docs as Person[]).map((person) => new MeiliPersonEntityV1_4({}).fromMongoDocument(person));
             break;
           default:
             throw new Error(`Unknown entity type: ${entityType}`);
@@ -99,6 +105,7 @@ export class SearchSyncService implements OnModuleInit {
   async syncMovies() {
     this.logger.log('Starting sync for movies');
     await this.syncEntity<Movie>(MOVIE_INDEX, this.movieService, 1000);
+    await this.syncEntity<Movie>(MOVIE_V1_4_INDEX, this.movieService, 1000);
     this.logger.log('Finished sync for movies');
   }
 
@@ -106,6 +113,7 @@ export class SearchSyncService implements OnModuleInit {
   async syncPersons() {
     this.logger.log('Starting sync for persons');
     await this.syncEntity<Person>(PERSON_INDEX, this.personService, 1000);
+    await this.syncEntity<Person>(PERSON_V1_4_INDEX, this.personService, 1000);
     this.logger.log('Finished sync for persons');
   }